src/third_party/blink/web_tests/external/wpt/css/css-scroll-snap/scroll-snap-type-on-root-element-expected.txt

This is a testharness.js-based test.
FAIL The scroll-snap-type on the root element is applied assert_equals: expected 515 but got 1000
FAIL The writing-mode on the body is used assert_equals: expected 515 but got 1000
Harness: the test ran to completion.
